requirements.txt
wheel
flask
python-dotenv
mongoengine
mongomock
PyJWT
flask_limiter